requirements.txt
Django>=4.2,<5.0
Pillow>=10.0
whitenoise>=6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
groq>=0.9
python-dotenv>=1.0
